+++ src/main/java/com/game/query/controller/NumberStatisticsQueryViewController.java
@@ -3,23 +3,27 @@ package com.game.query.controller;
 import com.game.query.dto.NumberStatisticsResponse;
 import com.game.query.service.NumberStatisticsQueryService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-@RestController
+@Controller
 @RequiredArgsConstructor
-public class NumberStatisticsQueryController {
+public class NumberStatisticsQueryViewController {
 
     private final NumberStatisticsQueryService numberStatisticsQueryService;
 
     @GetMapping("/number-statistics/{level}")
-    public ResponseEntity<List<NumberStatisticsResponse>> getNumberStatistics(@PathVariable String level) {
-        List<NumberStatisticsResponse> response = numberStatisticsQueryService.findByLevel(level);
-        return ResponseEntity.ok(response);
+    public String getGameRank(@PathVariable String level, Model model) {
+        List<NumberStatisticsResponse> numberStatistics = numberStatisticsQueryService.findByLevel(level);
+
+        model.addAttribute("level", level);
+        model.addAttribute("numberStatistics", numberStatistics);
+
+        return "number-statistics";
     }
 
 }
