+++ src/main/resources/templates/home.html
@@ -14,15 +14,75 @@
 
         body {
             background: linear-gradient(to right, #667eea, #764ba2);
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
             color: white;
+            min-height: 100vh;
+        }
+
+        .navbar {
+            background-color: rgba(0, 0, 0, 0.2);
+            backdrop-filter: blur(8px);
+            padding: 1rem 2rem;
+            display: flex;
+            gap: 2rem;
+            position: sticky;
+            top: 0;
+            z-index: 1000;
+        }
+
+        .menu-item {
+            position: relative;
+        }
+
+        .menu-link {
+            color: white;
+            text-decoration: none;
+            font-weight: bold;
+            padding: 0.5rem 1rem;
+            border-radius: 8px;
+            transition: background-color 0.3s;
+            cursor: pointer;
+            display: inline-block;
+        }
+
+        .menu-link:hover {
+            background-color: rgba(255, 255, 255, 0.2);
+        }
+
+        .submenu {
+            display: none;
+            position: absolute;
+            top: 100%;
+            left: 0;
+            background-color: rgba(0, 0, 0, 0.3);
+            border-radius: 10px;
+            padding: 0.5rem 0;
+            min-width: 200px;
+        }
+
+        .submenu a {
+            display: block;
+            padding: 0.5rem 1rem;
+            color: white;
+            text-decoration: none;
+            transition: background-color 0.3s;
+        }
+
+        .submenu a:hover {
+            background-color: rgba(255, 255, 255, 0.2);
+        }
+
+        .menu-item:hover .submenu {
+            display: block;
         }
 
         .container {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
             text-align: center;
+            padding: 4rem 2rem;
+            margin-top: 120px; /* 이 부분을 추가하여 상단 여백을 늘려줍니다. */
         }
 
         h1 {
@@ -68,10 +128,42 @@
                 flex-direction: column;
                 align-items: center;
             }
+
+            .navbar {
+                flex-direction: column;
+                gap: 1rem;
+            }
+
+            .submenu {
+                position: static;
+            }
         }
     </style>
 </head>
 <body>
+
+<nav class="navbar">
+    <div class="menu-item">
+        <span class="menu-link">난이도 별 생성된 숫자 빈도 수</span>
+        <div class="submenu">
+            <a href="/number-statistics/easy">쉬움</a>
+            <a href="/number-statistics/normal">보통</a>
+            <a href="/number-statistics/hard">어려움</a>
+        </div>
+    </div>
+    <!-- 추후 추가 가능 -->
+    <!--
+    <div class="menu-item">
+        <span class="menu-link">숫자 맞춘 기록 랭킹</span>
+        <div class="submenu">
+            <a href="/ranking/easy">쉬움</a>
+            <a href="/ranking/normal">보통</a>
+            <a href="/ranking/hard">어려움</a>
+        </div>
+    </div>
+    -->
+</nav>
+
 <div class="container">
     <h1>난이도를 선택하세요</h1>
     <div class="difficulty-options">
@@ -92,9 +184,9 @@
 
 <script>
     function startGame(level) {
-        // 경로 이동: /easy, /normal, /hard
         location.href = `/game/${level}`;
     }
 </script>
+
 </body>
 </html>
